requirements.txt
streamlit==1.39.0
torch==2.4.0
torchvision==0.19.0
Pillow==10.4.0
numpy==1.26.4
matplotlib==3.9.2
